[Decomp][Player] Cohesion split: extract item validation methods to PlayerItemValidation.cpp
